Refactor: BlindIndex annotation, SQLAlchemy folder split, normalization, README rewrite
- Add non-SQLAlchemy BlindIndex annotation with per-field method selection
- Add blind index adapters (HMAC-SHA256, Argon2) in adapters/blind_index/
- Add normalization utility with strip/normalize options for blind indexes
- Split integrations/sqlalchemy.py into sqlalchemy/ package (encryption, hashing, blind_index)
- Add normalization support to SQLAlchemyBlindIndexValue
- Remove Fernet as default ENCRYPTION_METHOD (must be explicitly set)
- Restructure tests: split test_sqlalchemy_type.py into test_sqlalchemy/ folder
- Add tests for normalization and BlindIndex annotation
- Rewrite README: shorter, pure SQLAlchemy examples, no sqlmodel references
