+++ .github/workflows/run_tests.yml
@@ -28,6 +28,7 @@ jobs:
           poetry run pytest --cov=pydantic_encryption --cov-branch --cov-report=xml
         env:
           ENCRYPTION_KEY: ${{ secrets.ENCRYPTION_KEY }}
+          ENCRYPTION_METHOD: "fernet"
           BLIND_INDEX_SECRET_KEY: "test-blind-index-secret-key"
 
       - name: Upload coverage reports to Codecov
