+++ pydantic_encryption/adapters/blind_index/__init__.py
@@ -1,0 +1,8 @@
+from pydantic_encryption.adapters.blind_index import hmac_sha256
+
+from pydantic_encryption.adapters.blind_index.hmac_sha256 import HMACSHA256Adapter
+
+__all__ = [
+    "hmac_sha256",
+    "HMACSHA256Adapter",
+]
