+++ pydantic_encryption/integrations/sqlalchemy/hashing.py
@@ -1,0 +1,47 @@
+from pydantic_encryption._lazy import require_optional_dependency
+
+require_optional_dependency("sqlalchemy", "sqlalchemy")
+
+from sqlalchemy.types import LargeBinary, TypeDecorator
+
+from pydantic_encryption.adapters import hashing
+from pydantic_encryption.types import HashedValue
+
+
+class SQLAlchemyHashed(TypeDecorator):
+    """Type adapter for SQLAlchemy to hash strings using Argon2."""
+
+    impl = LargeBinary
+    cache_ok = True
+
+    def process_bind_param(self, value: str | bytes | None, dialect) -> bytes | None:
+        """Hashes a string before binding it to the database."""
+
+        if value is None:
+            return None
+
+        return hashing.argon2.Argon2Adapter.hash(value)
+
+    def process_literal_param(self, value: str | bytes | None, dialect) -> HashedValue | None:
+        """Hashes a string for literal SQL expressions."""
+
+        if value is None:
+            return None
+
+        processed = hashing.argon2.Argon2Adapter.hash(value)
+
+        return dialect.literal_processor(self.impl)(processed)
+
+    def process_result_value(self, value: str | bytes | None, dialect) -> HashedValue | None:
+        """Returns the hash value as-is from the database, wrapped as a HashedValue."""
+
+        if value is None:
+            return None
+
+        return HashedValue(value)
+
+    @property
+    def python_type(self):
+        """Return the Python type this is bound to (str)."""
+
+        return self.impl.python_type
